:bug: Change SITEURL by main_siteurl (fix #29)
This commit change the variable SITEURL in the links to the assets in the head of base.html (<link rel>) by mail_siteurl because if in a multilingual website is used SITEURL, this one change for each language, so it is necessary a variable that doesn't change across the language options. See #29 for more information about it.